Removed compiler warnings

+++ tests/generate-tests.cpp
@@ -14,8 +14,8 @@ TEST_CASE("Generate changing and creating elements", "[Generate][Basic functiona
 
     SECTION("Should be 0, 1, 2, 3") {
         size_t expected = 0;
-        for (int i : generator) {
-            CHECK((size_t)i == expected++);
+        for (size_t i : generator) {
+            CHECK(i == expected++);
         }
     }
 }
